# apps/server/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/payments": {
      "get": {
        "operationId": "listPayments",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentListResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/payments/{id}": {
      "get": {
        "operationId": "getPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/projects/{id}/payments": {
      "post": {
        "operationId": "createProjectPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/customers/{id}/payments": {
      "post": {
        "operationId": "createCustomerPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/payments/{id}/actions/verify": {
      "post": {
        "operationId": "verifyPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/payments/{id}/actions/settle": {
      "post": {
        "operationId": "settlePayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/payments/{id}/actions/reverse": {
      "post": {
        "operationId": "reversePayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialReversalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/payments/{id}/allocations": {
      "post": {
        "operationId": "allocatePayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyValueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/payment-allocations/{id}/actions/reverse": {
      "post": {
        "operationId": "reversePaymentAllocation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialReversalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/payment-allocations/{id}/actions/create-deposit-balance": {
      "post": {
        "operationId": "createDepositBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepositBalanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositBalanceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/projects/{id}/deposit-balances": {
      "get": {
        "operationId": "listProjectDepositBalances",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositBalanceListResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/deposit-balances/{id}/applications": {
      "post": {
        "operationId": "applyDepositBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyValueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositBalanceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/deposit-applications/{id}/actions/reverse": {
      "post": {
        "operationId": "reverseDepositApplication",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialReversalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositBalanceDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/customers/{id}/customer-credits": {
      "get": {
        "operationId": "listCustomerCredits",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerCreditListResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      },
      "post": {
        "operationId": "createCustomerCredit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerCreditDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerCreditDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/customer-credits/{id}/applications": {
      "post": {
        "operationId": "applyCustomerCredit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyValueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerCreditDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/customer-credit-applications/{id}/actions/reverse": {
      "post": {
        "operationId": "reverseCustomerCreditApplication",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialReversalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerCreditDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
...
  },
  "info": {
    "title": "Lincoln Dirt and Gravel API",
    "description": "Lincoln Dirt and Gravel operating system API",
    "version": "1.7.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "paymentNumber": {
            "type": "string"
          },
          "customerAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "customerName": {
            "type": "string"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "projectNumber": {
            "type": "string",
            "nullable": true
          },
          "amountCents": {
            "type": "number"
          },
          "allocatedCents": {
            "type": "number"
          },
          "customerCreditCents": {
            "type": "number"
          },
          "refundedCents": {
            "type": "number"
          },
          "availableCents": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "receivingAccountReference": {
            "type": "string"
          },
          "providerName": {
            "type": "string",
            "nullable": true
          },
          "providerTransactionId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "payer": {
            "type": "object",
            "additionalProperties": true
          },
          "receiptStatus": {
            "type": "string"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "verifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "settledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reversedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reversalReason": {
            "type": "string",
            "nullable": true
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValueApplicationDto"
            }
          }
        },
        "required": [
          "id",
          "paymentNumber",
          "customerAccountId",
          "customerName",
          "amountCents",
          "allocatedCents",
          "customerCreditCents",
          "refundedCents",
          "availableCents",
          "currency",
          "paymentMethod",
          "receivingAccountReference",
          "status",
          "payer",
          "receiptStatus",
          "receivedAt",
          "allocations"
        ]
      },
      "PaymentListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentDto"
            }
          }
        },
        "required": ["items"]
      },
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "amountCents": {
            "type": "number",
            "minimum": 1
          },
          "currency": {
            "type": "string",
            "default": "USD",
            "pattern": "^[A-Z]{3}$"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "cash",
              "zelle",
              "venmo",
              "cash_app",
              "paypal",
              "card",
              "bank_transfer",
              "check"
            ]
          },
          "receivingAccountReference": {
            "type": "string",
            "maxLength": 160
          },
          "payerName": {
            "type": "string",
            "maxLength": 200
          },
          "payerEmail": {
            "type": "string",
            "maxLength": 320
          },
          "providerName": {
            "type": "string",
            "maxLength": 80
          },
          "providerTransactionId": {
            "type": "string",
            "maxLength": 200
          },
          "evidenceDocumentId": {
            "type": "string",
            "format": "uuid"
          },
          "receiptStatus": {
            "type": "string",
            "enum": ["missing", "attached", "issued", "waived", "not_required"]
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["amountCents", "paymentMethod", "receivingAccountReference", "payerName"]
      },
      "FinancialReversalDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": ["reason"]
      },
      "ApplyValueDto": {
        "type": "object",
        "properties": {
          "invoiceId": {
            "type": "string",
            "format": "uuid"
          },
          "amountCents": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": ["invoiceId", "amountCents"]
      },
      "CreateDepositBalanceDto": {
        "type": "object",
        "properties": {
          "depositType": {
            "type": "string",
            "default": "advance_payment",
            "enum": ["advance_payment", "split_advance"]
          }
        }
      },
...
      "DepositBalanceListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DepositBalanceDto"
            }
          }
        },
        "required": ["items"]
      },
...
      "CustomerCreditListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerCreditDto"
            }
          }
        },
        "required": ["items"]
      },
...
    }
  }
}